refactor: rename header.js to index.js
